backend/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
python-dotenv==1.0.1
requests==2.32.3